libc: Move on_exit and __cxa_atexit in lib_atexit.c
